Extract PayPalWebView from PayPalWebFragment.
